create axiosWithAuth file to assign token to localstorage and renamed some files
